vt100.jspp: Disable insane handling of events in keyDown
Depending on your keyboard layout and browser, the keyCode might not be
what you expect (e.g. Firefox with an azerty layout returns 170 for
'*'). This seems impossible to handle correctly, so only rely on
keyPressed. This means that some shortcuts are interpreted by the
browser itself (e.g. Ctrl-K with Chrome), too bad.